requirements.txt
absl-py==0.15.0
argon2-cffi==21.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.8
clang==5.0
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cycler==0.10.0
Cython==0.29.24
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
descartes==1.1.0
efficientnet-pytorch==0.6.3
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.4.0
Fiona==1.8.20
flatbuffers==1.12
fuzzywuzzy==0.18.0
gast==0.4.0
geopandas==0.9.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.1.0
huggingface-hub==0.2.1
idna==3.3
ijson==3.1.4
importlib-metadata==4.8.1
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.17.2
Jinja2==3.0.2
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
munch==2.5.0
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.4
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.4.60
openpyxl==3.0.9
opt-einsum==3.3.0
osmnx==0.16.2
packaging==21.0
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pkg_resources==0.0.0
pretrainedmodels==0.7.4
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyproj==3.0.1
pyrsistent==0.18.0
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
Rtree==0.9.7
sacremoses==0.0.46
scikit-learn==0.24.2
scipy==1.5.4
segmentation-models-pytorch==0.2.1
Send2Trash==1.8.0
Shapely==1.8.0
signal-envelope==1.3
six==1.15.0
sklearn==0.0
spectrum==0.8.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
timm==0.4.12
tokenizers==0.10.3
torch==1.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
transformers==4.13.0
typing-extensions==3.7.4.3
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
zipp==3.6.0
